consolidate machineGroups code in two files
