+++ inductiva/resources/__init__.py
@@ -1,6 +1,3 @@
 #pylint: disable=missing-module-docstring
-from .machines import MachineGroup, ElasticMachineGroup, MPICluster
-from .machine_types import get_available_machine_types
-from .machine_groups import estimate_machine_cost
-from .machine_types import list_available_machines
-from .machine_groups import get
+from .machine_groups import MachineGroup, ElasticMachineGroup, MPICluster
+from .utils import estimate_machine_cost, get, get_available_machine_types, get_by_name, get_machine_dict, list_available_machines
