+++ inductiva/tests/resources/test_machine_types.py
@@ -32,7 +32,7 @@ RESPONSE = [{
 
 def test_get_available_machine_types():
     mock_compute_api_path =\
-        "inductiva.resources.machine_types.compute_api.ComputeApi"
+        "inductiva.resources.utils.compute_api.ComputeApi"
     with mock.patch(mock_compute_api_path) as mock_compute_api:
         mock_inner_response = mock.MagicMock()
         mock_inner_response.data = mock.MagicMock()
@@ -43,7 +43,7 @@ def test_get_available_machine_types():
             return_value=mock_response)
         (mock_compute_api.return_value.list_available_machine_types
         ) = mock_list_available_machine_types
-        result = inductiva.resources.machine_types.get_available_machine_types(
+        result = inductiva.resources.get_available_machine_types(
             provider="GCP")
         assert result[0]["machine_type"] == "c2-standard-4"
         assert result[0]["num_cpus"] == 4
